move test files to match new module locations
