+++ packages/tracing/test/integrations/node/postgres.test.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/unbound-method */
 import { Hub, Scope } from '@sentry/hub';
 
-import { Postgres } from '../../src/integrations/node/postgres';
-import { Span } from '../../src/span';
+import { Postgres } from '../../../src/integrations/node/postgres';
+import { Span } from '../../../src/span';
 
 class PgClient {
   // https://node-postgres.com/api/client#clientquery
